Usuarios

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -9,9 +10,14 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $users = \App\Models\User::all();
+        $users = \App\Models\User::query();
+        $users->when($request->filled('search'), function($query) use ($request) {
+            $query->where('name', 'like', '%' . $request->search . '%')
+                ->orWhere('email', 'like', '%' . $request->search . '%');
+        });
+        $users = $users->paginate();
         return view('users.index', compact('users'));
     }
 
@@ -54,6 +60,7 @@ class UserController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('edit', \App\Models\User::class);
         $user = \App\Models\User::findOrFail($id)
            ->load(['perfil', 'interests']);
         $roles = \App\Models\Role::all();
@@ -86,6 +93,7 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('destroy', \App\Models\User::class);
         $user = \App\Models\User::findOrFail($id);
         $user->delete();
         return redirect()
